# transformers/bike-boxes/documentation/loewenbytes/260430_openapi-suedtirolmobile-bikebox-api-loewenbytes-v1.0.1.yaml
...
openapi: 3.0.3
info:
  title: Loewenbytes Bikebox Provider API for Suedtirolmobile App
  version: 1.0.1
  contact:
    email: info@loewenbytes.com
  description: |
    # Welcome to the Loewenbytes Bikebox Provider API
    
    This API provides functionality for integrating with and managing Bikebox Provider services.
    
    ## Common notes
    
    ### Encoding
    
    Unless described in other ways, all requests MUST use, and all response WILL
    use, characters that are encoded in UTF-8.
    
    ### Localization
    
    This API provides localized text for DE, EN, IT, LAD.
    
    ### Common (global) Request Headers
    
    Not required.
    
    ### Common Error handling
    
    Unless otherwise noted, each API error response includes a X-ErrID Response
    Header, with an error id, that could be displayed to consumers for further
    error analysis.
    
    Additionally, each error response MAY contain a common error object $ref:
    '#/components/schemas/ErrorResponse'
    
    #### Response Status Error Codes
    
    Each API may response one of the standard HTTP Status Codes, even if not
    described for each API in detail:
    
    * 400 Bad Request – client sent an invalid request, such as lacking
    required request body or parameter or more conditions in the request data
    evaluated to false
    * 500 Internal Server Error – a generic error occurred on the server
    * 503 Server Unavailable Error –  server is currently unable to handle the request
...
servers:
...
paths:
...
  /auth/refreshToken:
    post:
      summary: Refresh a Customer Token.
      description: |
        Use this endpoint to refresh a **Customer Token** issued from a previous customer sign-in via `/customers/signIn` or `/customers/signInMifareId`.
        
        **Note:** This endpoint requires sending the `refreshToken` as a body parameter instead of using the defined authorization header. This behavior may be adjusted in future versions if necessary.
      operationId: refreshToken
      tags:
        - Authentication
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/CustomerRefreshTokenRequest'
      responses:
        200:
          description: Returns the requested data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerTokenResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

      # /auth/.well-known/jwks.json defined wihtin LWS scope
      # get:
      #    summary: Get JSON Web Key Set (JWKS, see https://datatracker.ietf.org/doc/html/rfc7517) to verify tokens, that was created by this APIs issuer.
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        errorCode:
          description: |
            Well-known error codes as defined in "Bike Box API Integration in südtirolmobil App - From app code" provided by Runelab Srl/Marco Brugali.
            
            **Note:** Additional error codes may be added on demand.
          type: string
          enum:
            - undefined
            - hardwareUnlockError
            - qrCodeNotValid
            - reserveNotInProgress
            - notFound
            - placesNotAvailable
            - reservationAlreadystarted
          readOnly: true
        errorMessage:
          type: string
          readOnly: true
      required:
        - errorCode
        - errorMessage
...
    CustomerRefreshTokenRequest:
      description: Token request data, according to https://datatracker.ietf.org/doc/html/rfc6749#section-4.3.2
      type: object
      properties:
        grant_type:
          description: Value MUST be set to "refresh_token", see https://datatracker.ietf.org/doc/html/rfc6749#section-6
          type: string
          enum: [ refresh_token ]
        refresh_token:
          description: A refresh token issued to the client
          type: string
          minLength: 5
      required:
        - grant_type
        - refresh_token
    CustomerTokenResponse:
      type: object
      properties:
        tokenID:
          type: string
          example: eyJhbGciOiJSUIsImtpZCI6IjNmNjcyNeyJhbGciOiJSUIsImtpZCI6IjN=
        refreshToken:
          type: string
          example: eyJhbGciOiJSUIsImtpZCI6IjNmNjcyNeyJhbGciOiJSUIsImtpZCI6IjN=
      required:
        - tokenID
        - refreshToken
  securitySchemes:
    AccessToken:
      type: oauth2
      description: |
        Obtain an **Access Token** via the OAuth 2.0 Client Credentials flow.
        Contact **[info@loewenbytes.com](mailto:info@loewenbytes.com)** for `client_id` and `client_secret`.
      flows:
        clientCredentials:
          tokenUrl: https://test.app.loewenbytes.com/api/v1/oauth/token
          scopes: { }

    CustomerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        **Customer Token** for authenticated customers or end-users.
        Can be obtained via `/customers/signIn` or `/customers/signInMifareId`.